@@ -12,47 +12,47 @@ packages = [
     { include = "{{ cookiecutter.project_slug_no_hyphen }}" },
 ]
 classifiers = [
-    "Programming Language :: Python :: 3.6",
     "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
 ]
 
 
 [tool.poetry.dependencies]
-python = "^3.6.1"
-click = "^8.0.0"
-rich = "^10.3.0"
-PyYAML = "^5.4.1"
+python = "^3.7.0"
+click = ">=8.0.0"
+rich = ">=10.3.0"
+PyYAML = ">=5.4.1"
 
 [tool.poetry.dev-dependencies]
-pytest = "^6.2.3"
-coverage = {extras = ["toml"], version = "^5.3"}
-safety = "^1.9.0"
-typeguard = "^2.12.0"
-xdoctest = {extras = ["colors"], version = "^0.15.0"}
-sphinx = "^4.0.2"
-sphinx-autobuild = "^2021.3.14"
-pre-commit = "^2.11.1"
-flake8 = "^3.8.4"
-black = "^20.8b1"
-flake8-bandit = "^2.1.2"
-flake8-bugbear = "^21.4.3"
-flake8-docstrings = "^1.5.0"
-flake8-rst-docstrings = "^0.2.3"
-pep8-naming = "^0.11.1"
-darglint = "^1.5.8"
-reorder-python-imports = "^2.5.0"
-pre-commit-hooks = "^4.0.1"
-sphinx-rtd-theme = "^0.5.0"
-sphinx-click = "^3.0.0"
-Pygments = "^2.8.1"
-types-pkg-resources = "^0.1.2"
-types-requests = "^2.25.2"
-types-attrs = "^19.1.0"
-sphinx-rtd-dark-mode = "^1.2.3"
-Jinja2 = "^3.0.1"
-mypy = "^0.910"
+pytest = ">=6.2.3"
+coverage = {extras = ["toml"], version = ">=5.3"}
+safety = ">=1.9.0"
+typeguard = ">=2.12.0"
+xdoctest = {extras = ["colors"], version = ">=0.15.0"}
+sphinx = ">=4.0.2"
+sphinx-autobuild = ">=2021.3.14"
+pre-commit = ">=2.11.1"
+flake8 = ">=3.8.4"
+black = ">=20.8b1"
+flake8-bandit = ">=2.1.2"
+flake8-bugbear = ">=21.4.3"
+flake8-docstrings = ">=1.5.0"
+flake8-rst-docstrings = ">=0.2.3"
+pep8-naming = ">=0.11.1"
+darglint = ">=1.5.8"
+pre-commit-hooks = ">=4.0.1"
+sphinx-rtd-theme = ">=0.5.0"
+sphinx-click = ">=3.0.0"
+Pygments = ">=2.8.1"
+types-pkg-resources = ">=0.1.2"
+types-requests = ">=2.25.2"
+types-attrs = ">=19.1.0"
+sphinx-rtd-dark-mode = ">=1.2.3"
+Jinja2 = ">=3.0.1"
+mypy = ">=0.910"
+pyupgrade = ">=2.31.0"
 
 [tool.poetry.scripts]
 {{ cookiecutter.project_slug }} = "{{ cookiecutter.project_slug_no_hyphen }}.__main__:main"
